moved footer, header and nav components to the component folder and renamed them all to have first letter capital
