+++ client/src/components/Header.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Nav from "./nav";
+import Nav from "./Nav";
 
 function Header(props) {
   return (
