+++ client/src/pages/Home.js
@@ -1,8 +1,8 @@
 // import { useEffect, useState } from "react";
 // import { Link } from "react-router-dom";
-import Header from "../../header.js";
+import Header from "../components/Header.js";
 import "./home.css";
-import Footer from "../footer.js";
+import Footer from "../components/Footer.js";
 import VideoHero from "../components/VideoHero.js";
